% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cr_fbm.R
\name{calc_cr_fbm}
\alias{calc_cr_fbm}
\title{Calculate root reinforcement at current strain according to FBM model}
\usage{
calc_cr_fbm(
  epsr0rel,
  dr,
  phir,
  tru,
  betaF,
  tru0,
  k = 1.2,
  dr0 = 1,
  sumoutput = TRUE
)
}
\arguments{
\item{epsr0rel}{tensile strain in reference root, normalised by the tensile
strain to peak in this root (numeric array, size n)}

\item{dr}{representative diameter of root class (numeric array, size m)}

\item{phir}{root area ratio of roots in class (numeric array, size m)}

\item{tru}{tensile strength of roots in class (numeric array, size m)}

\item{betaF}{load sharing parameter (numeric scalar)}

\item{tru0}{Tensile strength of root with reference diameter (numeric array)}

\item{k}{Wu/Waldron factor accounting for root orientations at failure
(numeric scalar, default 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}

\item{sumoutput}{If TRUE, a total reinforcement is outputted for every
strain step. If FALSE, a reinforcement matrix is outputted
(rows: strain step, columns: classes)}
}
\value{
numeric array with root reinforcement preductions `cr_fbm`
}
\description{
This function calculates the current root reinforcement according to the
FBM, assuming discrete root classes, given the current level of (normalised)
strain in the reference root
}
\examples{
calc_cr_fbm(seq(0,2.5,l=101), seq(1,4), rep(0.01,4), 10e3*seq(4)^-0.5, 1, 10e3)
}
