% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cru_fbm.R
\name{calc_cru_fbm}
\alias{calc_cru_fbm}
\title{Calculate peak root reinforcement according to FBM model}
\usage{
calc_cru_fbm(dr, phir, tru, betaF, tru0, k = 1.2, dr0 = 1)
}
\arguments{
\item{dr}{representative diameter of root class (numeric array, size m)}

\item{phir}{root area ratio of roots in class (numeric array, size m)}

\item{tru}{tensile strength of roots in class (numeric array, size m)}

\item{betaF}{load sharing parameter (numeric scalar)}

\item{tru0}{Tensile strength of root with reference diameter (numeric scalar)}

\item{k}{Wu/Waldron factor accounting for root orientations at failure
(numeric scalar, default 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
numeric scalar with root reinforcement preductions `cru_fbm`
}
\description{
This function calculates the peak root reinforcement according to the
FBM, assuming discrete root classes
}
\examples{
calc_cru_fbm(seq(1,4), rep(0.01,4), 10e3*seq(4)^-0.5, 1, 10e3)
}
