% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cru_fbmw.R
\name{calc_cru_fbmw}
\alias{calc_cru_fbmw}
\title{Calculate peak root reinforcement according to FBMw model}
\usage{
calc_cru_fbmw(
  dr,
  phir,
  tru,
  betaF,
  tru0,
  kappa,
  k = 1.2,
  dr0 = 1,
  optim_method = "BFGS"
)
}
\arguments{
\item{dr}{representative diameter of root classes (numeric array, size m)}

\item{phir}{root area ratio of roots in classes (numeric array, size m)}

\item{tru}{tensile strength of roots in classes (numeric array, size m)}

\item{betaF}{load sharing parameter (numeric scalar)}

\item{tru0}{Tensile strength of root with reference diameter (numeric scalar)}

\item{kappa}{Weibull shape parameter required for the root survival
function (scalar)}

\item{k}{Wu/Waldron factor accounting for root orientations at failure
(numeric scalar, default 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}

\item{optim_method}{Method used to find peak in optimalisation procedure.
`optim_method='BFGS' uses the `optim` function and makes an informed guess
for the starting value. `optim_method='Brent'` uses the `optimize` function
and seaches within bounds. For high values of `kappa`, the latter method
does not always find the global maximum if located near the edges of the
search domain, probably because of 'sawtooth' nature of the trace, and
therefore `optim_method='BFGS'` is the default.}
}
\value{
numeric scalar with peak root reinforcement preductions `cru_fbmw`
}
\description{
This function calculates the peak root reinforcement according to the
FBMw, assuming discrete root classes.
}
\examples{
calc_cru_fbmw(seq(1,4), rep(0.01,4), 10e3*seq(4)^-0.5, 1, 10e3, 4)
}
