% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kk_fbmc.R
\name{calc_kk_fbmc}
\alias{calc_kk_fbmc}
\title{Calculate root reinforcement at current strain according to FBMc model}
\usage{
calc_kk_fbmc(epsr0rel, drmin, drmax, betaF, betat, betaphi, dr0 = 1)
}
\arguments{
\item{epsr0rel}{tensile strain in reference root, normalised by the tensile
strain to peak in this root (numeric array)}

\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaF}{Load sharing coefficient (numeric array)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
numeric array with peak reinforcement preductions `cr_fbmc`
}
\description{
This function calculates the current root reinforcement, given a strain
according to the FBMc model. The results are normalised by the corresponding
WWMc model predictions, resulting in a k" coefficient.
}
\examples{
calc_kk_fbmc(seq(0,1,l=25), 1, 5, 1, -0.5, -0.2)
}
