% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kk_fbmcw.R
\name{calc_kk_fbmcw}
\alias{calc_kk_fbmcw}
\title{Calculate root reinforcement at current strain according to FBMcw model}
\usage{
calc_kk_fbmcw(epsr0rel, drmin, drmax, betaF, betat, betaphi, kappa, dr0 = 1)
}
\arguments{
\item{epsr0rel}{tensile strain in reference root, normalised by the tensile
strain to peak in this root (numeric array)}

\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaF}{Load sharing coefficient (numeric array)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{kappa}{Weibull shape parameter for the root survival function
(numeric array)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
numeric array with peak reinforcement preductions `cr_fbmcw`
}
\description{
This function calculates the current root reinforcement, given a strain
according to the FBMcw model. The results are normalised by the corresponding
WWMc model predictions, resulting in a k" coefficient.
}
\examples{
calc_kk_fbmc(seq(0,2,l=50), 1, 5, 1, -0.5, -0.2, 2.5)
}
