% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kku_fbmc.R
\name{calc_kku_fbmc}
\alias{calc_kku_fbmc}
\title{Calculate sequential mobilisation reduction factor k" for FBMc model}
\usage{
calc_kku_fbmc(drmin, drmax, betaF, betat, betaphi)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaF}{load sharing coefficient (numeric array)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}
}
\value{
numeric array with reduction coefficients k" (numeric array)
}
\description{
This function calculates reduction in peak root reinforcement due to
sequential mobilisation of roots according to the FBMc model, compared
to the reinforcement predicted by the WWMc model
}
\examples{
calc_kku_fbmc(1, 5, 1, -0.5, -1)
}
