% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kku_fbmcw.R
\name{calc_kku_fbmcw}
\alias{calc_kku_fbmcw}
\title{Calculate sequential mobilisation reduction factor k" for FBMcw model}
\usage{
calc_kku_fbmcw(drmin, drmax, betaF, betat, betaphi, kappa, dr0 = 1)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaF}{load sharing coefficient (numeric array)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{kappa}{Weibull shape parameter for the root survival function
(numeric array)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
numeric array with reduction coefficients k" (numeric array)
}
\description{
This function calculates reduction in peak root reinforcement due to
sequential mobilisation of roots according to the FBMcw model, compared
to the reinforcement predicted by the WWMc model
The peak reinforcement is found using the `optimize` function in base-R
}
\examples{
calc_kku_fbmcw(1, 5, 1, -0.5, -1, 2.5)
}
