% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phir0.R
\name{calc_phir0}
\alias{calc_phir0}
\title{Calculate root area ratio at reference diameter}
\usage{
calc_phir0(phirt, drmin, drmax, betaphi, dr0 = 1)
}
\arguments{
\item{phirt}{Total root area ratio in bundle (numeric array)}

\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
numeric array with `phir0` values
}
\description{
This function calculated the root area ratio at the reference diameter
based on the total root area ratio, the diameter range and a power-law
coefficient describing the root bundle.
}
\examples{
calc_phir0(0.01, 1, 5, -1)
}
