% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cumulative_phir.R
\name{create_cumulative_phir}
\alias{create_cumulative_phir}
\title{Create a diameter versus cumulative root area ratio trace}
\usage{
create_cumulative_phir(dclass)
}
\arguments{
\item{dclass}{dataframe with root area ratio data for classes. Should
contain fields for minimum (`drmin`) and maximum diameter in class
(`drmax`), and the root area ratio for each class (`phir`)}
}
\value{
dataframe with cumulative traces. Contain fields for root
  diameter (`dr`) and cumulative root area ratio (`phir_cum`)
}
\description{
This function creates a trace for root diameter versus cumulative root
area ratio, based on binned root diameter/root area ratio input
}
\examples{
dclass <- data.frame(
  drmin = c(1, 2, 3, 4),
  drmax = c(2, 3, 3.5, 5),
  phir = c(0.1, 0.2, 0.4, 0.7)
)
create_cumulative_phir(dclass)
}
