% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise_rootarearatio.R
\name{discretise_rootararatio}
\alias{discretise_rootararatio}
\title{Discretise root area ratio over a number of discrete diameter classes}
\usage{
discretise_rootararatio(drmin, drmax, betaphi, nc)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric scalars)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)#' @param phir root area ratio of roots in class
(numeric arra, size m)}

\item{nc}{The number of discrete, equal-width root classes to use}
}
\value{
a dataframe containing the fields `dr`, the diameter in the middle
  of each class, and `phirphirt`, the fraction of the total root area
  ratio that should be assigned to each class
}
\description{
This function splits a root diameter range into a number of equal-width
root diameter classes, and calculates the fraction on the total root
area ratio that should be assigned to each class. Function is superseded
by `discretise_rootclasses`.
}
\examples{
discretise_rootararatio(1, 5, -0.5, 10)
}
