% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise_rootclasses.R
\name{discretise_rootclasses}
\alias{discretise_rootclasses}
\title{Discretise root area ratio over a number of discrete diameter classes}
\usage{
discretise_rootclasses(drmin, drmax, tru0, betat, phirt, betaphi, nc, dr0 = 1)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric scalars)}

\item{tru0}{Tensile strength of root with reference diameter (numeric array)}

\item{betat}{Power coefficient for tensile strength (scalar)}

\item{phirt}{Total root area ratio (numeric scalar)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{nc}{The number of discrete, equal-width root classes to use}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
a dataframe containing the fields `dr`, the diameter in the middle
  of each class, `phir`, the root area ratio that should be assigned to
  each class, and `tru`, the tensile strength assigned to each class. It
  also return `drmini` and `drmaxi`, the lower and upper diameter in each class
}
\description{
This function splits a root diameter range into a number of equal-width
root diameter classes, and calculates the root area ratio that should
be assigned to each class as well as the tensile strength. Power law
distributions for root area ratio and tensile strength are assumed.
The representative diameter for each class is chosen halfway the range of
diameters in the class.
}
\examples{
discretise_rootclasses(1, 5, 10e3, -0.5, 0.01, -0.5, 10)
}
