% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_powerlaw_phir.R
\name{fit_powerlaw_phir}
\alias{fit_powerlaw_phir}
\title{Fit a bounded power-law trace to diameter versus cumulative root area ratio}
\usage{
fit_powerlaw_phir(dcum, drmin = 1e-12, drmax = Inf, betaphi_initial_guess = T)
}
\arguments{
\item{dcum}{dataframe with fields for the root diameter (`dr') and
cumulative root area ratio (`phir_cum`).}

\item{drmin}{an user-defined lower limit to the fitted
root diameter range. Should be larger than zero. Default 1e-12}

\item{drmax}{an user-defined upper limit to the fitted
root diameter range. Should be larger than zero. Default infinity}

\item{betaphi_initial_guess}{if TRUE, betaphi is first approximated
by fixing `drmin` and `drmax` between the known values in the input
this best fit is subsequently used in the 'real' fit. This should
increase the likelyhood of obtaining a suitable fitting result}
}
\value{
dataframe with fitting parameters. Returns:
  the total root area ratio (`phirt`)
  the best-fit power-law coefficient (`betaphi`)
  the best-fit minimum root diameter (`drmin`)
  the best-fit maximum root diameter (`drmax`)
  the Kolmogorov-Smirnov distance of the fit (`KS`)
}
\description{
This function fits a power-law curve to root diameter versus cumulative
root area ratio data. This power law is bounded by a minimum (`drmin`)
and maximum (`drmax`) root diameter.
}
\examples{
dcum <- data.frame(
  dr = c(1, 2, 3, 4, 5),
  phir_cum = c(0, 0.3, 0.5, 0.9, 0.95)
)
fit_powerlaw_phir(dcum)
}
