% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_loadsharingrange.R
\name{generate_loadsharingrange}
\alias{generate_loadsharingrange}
\title{Get range of load sharing parameters}
\usage{
generate_loadsharingrange(
  betaF = NA,
  betaFmin = -1,
  betaFmax = 3,
  n = 101,
  round = 0.2
)
}
\arguments{
\item{betaF}{Load sharing values that have to be included in the range
(numeric scalar)}

\item{betaFmin}{Minimum of the range (numeric scalar)}

\item{betaFmax}{Maximum of the range (numeric scalar)}

\item{n}{Number of discrete points used along the load sharing}

\item{round}{Round min and max of load sharing range to nearest multiple
of `round`}
}
\value{
a range with load sharing values betaF (numeric array)
}
\description{
This function generates a range of load sharing rules that can be
used for plotting
}
\examples{
generate_loadsharingrange(betaF = c(-4,1), betaFmin = -2, betaFmax = 2, n = 25)
}
