% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_loadsharingrules.R
\name{generate_loadsharingrules}
\alias{generate_loadsharingrules}
\title{Get load sharing parameters for common fibre bundle models}
\usage{
generate_loadsharingrules(
  betat = NULL,
  betaE = NULL,
  betaL = NULL,
  betaeps = NULL,
  betaF = NULL
)
}
\arguments{
\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric scalar)}

\item{betaE}{Power law coefficient for root diameter-root stiffness
fit (numeric scalar)}

\item{betaL}{Power law coefficient for root diameter-root length
fit (numeric scalar)}

\item{betaeps}{Power law coefficient for root diameter-root strain to peak
fit. This value overrules `betaE` if provided (numeric scalar)}

\item{betaF}{Additional user-defined load sharing rules (numerical array)}
}
\value{
dataframe with fields `loadsharingrule` and `betaF`, containing the name of the
  load sharing rule and the value of the load sharing coefficient respectively.
}
\description{
This function generates a list of load sharing rules associated
with common existing root reinforcement models. Rules are generated
based on the optional input parameters provided
}
\examples{
generate_loadsharingrules(betat = -0.5, betaE = 0, betaL = 0.7, betaF = 1.0)
}
