% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_strainrange.R
\name{generate_strainrange}
\alias{generate_strainrange}
\title{Generate suitable range of strains for plotting FBMc/FBMcw models}
\usage{
generate_strainrange(drmin, drmax, betaF, betat, dr0 = 1, margin = 1, n = 101)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric scalars)}

\item{betaF}{load sharing coefficient (numeric scalars)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric scalars)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}

\item{margin}{Multiplier to generate a larger strain range than required
for the FBMc model (numeric scalar, default 1.0)}

\item{n}{Number of discrete points used along the strain range}
}
\value{
array with strain of (normalised strain) `epsr0rel`
}
\description{
This function generates range of strains (defined as the strain in the
reference root, normalised by the strain to peak in the reference root)
that can be used for plotting the mobilisation of root reinforcement
This normalised strain is referred to as `epsr0rel`
}
\examples{
generate_strainrange(1, 5, 1, -0.5, margin = 1.5, n = 25)
}
