% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_strainrange_fbm.R
\name{generate_strainrange_fbm}
\alias{generate_strainrange_fbm}
\title{Generate range of strains for FBM model (discrete root classes)}
\usage{
generate_strainrange_fbm(
  dr,
  tru,
  betaF,
  tru0,
  includezero = T,
  includefailure = T,
  failureoffset = 1e-06,
  margin = NULL,
  dr0 = 1,
  outputperclass = F
)
}
\arguments{
\item{dr}{representative diameter of root class (numeric array, size m)}

\item{tru}{tensile strenght of roots in root class (numeric array, size m)}

\item{betaF}{load sharing parameter (numeric scalar)}

\item{tru0}{tensile strength of root with reference diameter}

\item{includezero}{If TRUE, includes a zero-strain point}

\item{includefailure}{if TRUE,  includes a extra strain point
just after failure in each class. The offset is governed by the
value of `failureoffset`}

\item{failureoffset}{The strain offset for strain points just after root
failure. This offset is equal to the value of `includefailure`
multiplied by strain to get the last root to fail}

\item{margin}{if specified, add an extra point at `margin*max(epsr0rel)'}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}

\item{outputperclass}{if TRUE, return the values of breakage strain for each input class}
}
\value{
numeric array with (reference) strains `epsr0rel`
}
\description{
This function generates a range of strains to be used for FBM calculations,
assuming discrete root classes. Strain is given as the (normalised)
strain in the reference root (tensile strain/tensile strain to peak). It
calculates the strains at failure of each root class, and optionally adds
a zero-strain points and the strain points just after root failures (to draw
nice 'sawtooth' plots).
}
\examples{
generate_strainrange_fbm(seq(1,5,l=10), 10e3*seq(1.5,1,l=10), 1, 10e3)
}
