% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_app_alltraces.R
\name{plotly_app_alltraces}
\alias{plotly_app_alltraces}
\title{Creates plotly for mobilisation of root reinforcement by all models}
\usage{
plotly_app_alltraces(dFBMc, cru_wwmc, dFBM, dFBMw, nsignif = 3)
}
\arguments{
\item{dFBMc}{dataframe with fields for the normalised reference strain
(`eps0rel`), the normalised reinforcement according to the FBMc model
(`kk_fbmc`) and the normalised reinforcement according to the FBMcw
model (`kk_fbmcw`).}

\item{cru_wwmc}{peak root-reinforcement according to the WWMc model
(numeric scalar)}

\item{dFBM}{dataframe with FBM model predictions. Contains fields
for the (normalised) reference strain `epsr0rel` and the corresponding
reinforcement `cr_fbm`}

\item{dFBMw}{dataframe with FBMw model predictions. Contains fields
for the (normalised) reference strain `epsr0rel` and the corresponding
reinforcement `cr_fbmw`}

\item{nsignif}{number of significant digits in plotly hover labels
(integer scalar)}
}
\value{
plotly object
}
\description{
This function generates a plotly bar plot showing how root
reinforcement mobilises according to the FBM, FBMw, FBMc and FBMcw
models, all in a single graph
}
\examples{
dFBMc <- data.frame(   #FBMc and FBMcw predictions
  epsr0rel = seq(0, 1, l = 51),
  kk_fbmc = seq(0, 1, l = 51)^2,
  kk_fbmcw = seq(0, 1, l = 51)^0.5
)
cru_wwmc <- 30      #WWMc prediction
dFBM <- data.frame(   #FBM predictions
  epsr0rel = seq(0, 1, l = 6),
  cr_fbm = 20 * seq(0, 1, l = 6)^2
)
dFBMw <- data.frame(   #FBMw predictions
  epsr0rel = seq(0, 1, l = 6),
  cr_fbmw = 15 * seq(0, 1, l = 6)^0.6
)
plotly_app_alltraces(dFBMc, cru_wwmc, dFBM, dFBMw, nsignif = 2)
}
