% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_app_cohesionpredictions.R
\name{plotly_app_cohesionpredictions}
\alias{plotly_app_cohesionpredictions}
\title{Creates plotly for peak reinforcements bar plot}
\usage{
plotly_app_cohesionpredictions(dm, cru_wwmc, nsignif = 3)
}
\arguments{
\item{dm}{dataframe with fields for existing load sharing parameters
(`betaF`), the name of this load sharing rule (`loadsharing'), the
normalised reinforcement according to the FBMc model (`kk_fbmc`) and
the normalised reinforcement according to the FBMcw model (`kk_fbmcw`).}

\item{cru_wwmc}{peak root-reinforcement according to the WWMc model
(numeric scalar)}

\item{nsignif}{number of significant digits in plotly hover labels
(integer scalar)}
}
\value{
plotly object
}
\description{
This function generates a plotly bar plot showing peak root
reinforcement according to the FBMc and FBMcw for a wide range
of existing load sharing models.
}
\examples{
drules <- data.frame(
  loadsharing = c('test1', 'test2'),
  betaF = c(0.5, 1),
  kk_fbmc = c(0.5, 1)^2,
  kk_fbmcw = c(0.5, 1)^0.5
)
cru_wwmc <- 30
plotly_app_cohesionpredictions(drules, cru_wwmc, nsignif = 2)
}
