% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_app_kkfbmc_kkfbmcw.R
\name{plotly_app_kkfbmc_kkfbmcw}
\alias{plotly_app_kkfbmc_kkfbmcw}
\title{Creates plotly for (normalised) mobilisation of FBMc and FBMcw models}
\usage{
plotly_app_kkfbmc_kkfbmcw(dt, nsignif = 3)
}
\arguments{
\item{dt}{dataframe with fields for the normalised reference strain
(`eps0rel`), the normalised reinforcement according to the FBMc model
(`kk_fbmc`) and the normalised reinforcement according to the FBMcw
model (`kk_fbmcw`).}

\item{nsignif}{number of significant digits in plotly hover labels
(integer scalar)}
}
\value{
plotly object
}
\description{
This function generates a plotly object showing how root reinforcement
is mobilised as function of reference strain for both the FBMc and FBMcw
models. Reinforcements are normalised by the WWMc solution. Strain is
defined as the current tensile strain in the reference root (dr=dr0)
normalised by the tensile strain to failure in the reference root.
}
\examples{
dt <- data.frame(
  epsr0rel = seq(0, 1, l = 51),
  kk_fbmc = seq(0, 1, l = 51)^2,
  kk_fbmcw = seq(0, 1, l = 51)^0.5
)
plotly_app_kkfbmc_kkfbmcw(dt)
}
