% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_stackedtrace_cr_fbm_fbmw.R
\name{plotly_stackedtrace_cr_fbm_fbmw}
\alias{plotly_stackedtrace_cr_fbm_fbmw}
\title{Plotly for normalised mobilisation of FBM reinforcement in discrete classes}
\usage{
plotly_stackedtrace_cr_fbm_fbmw(
  drmin,
  drmax,
  betaF,
  tru0,
  betat,
  phirt,
  betaphi,
  kappa,
  nc,
  k = 1.2,
  dr0 = 1,
  nround = 2,
  xlim = NULL,
  ylim = NULL,
  show_legend = T,
  hoverinfo_classes = "skip",
  nsignif = 3,
  plot_continuous = TRUE,
  plot_classes = TRUE
)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric scalar)}

\item{betaF}{Load sharing coefficient (numeric scalar)}

\item{tru0}{tensile strength of root with reference diameter (in kPa)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric scalar)}

\item{phirt}{Total root area ratio in bundle (numeric scalar)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric scalar)}

\item{kappa}{Weibull survival function shape parameter (numeric scalar)}

\item{nc}{The number of discrete, equal-width root classes to use}

\item{k}{Wu/Waldron factor accounting for root orientations at failure
(numeric scalar, default 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}

\item{nround}{number of decimals used in rounding diameter classes in plot
(default 2)}

\item{xlim}{optional max limit for x-axis}

\item{ylim}{optional max limit for y-axis}

\item{show_legend}{if TRUE, shows legend}

\item{hoverinfo_classes}{if TRUE, shows reinforcements for all root classes
in hoverinfo information. If `skip`, this information is not shown and
only information for the total reinforcement is displaced in the
hoverinfo}

\item{nsignif}{number of significant digits in plotly hover labels
(integer scalar)}

\item{plot_continuous}{if TRUE, also plot continuous solutions}

\item{plot_classes}{if TRUE, the effect contibution of individual classes is added}
}
\value{
Plotly object
}
\description{
Creates a plotly object showing how reinforcement is mobilised according
to the FBM model with a discrete number of equal-width root diameter
classes. Power-law relations for the root area ratio and tensile strength
are assumed. The results are normalised by the Wu/Waldron (WWM) solution
using the same classes.
}
\examples{
plotly_stackedtrace_cr_fbm_fbmw(1, 5, 1, 10e3, -0.5, 0.01, -0.2, 4, 5)
}
