% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_stackedtrace_kk_fbm_fbmw.R
\name{plotly_stackedtrace_kk_fbm_fbmw}
\alias{plotly_stackedtrace_kk_fbm_fbmw}
\title{Plotly for normalised mobilisation of FBM reinforcement in discrete classes}
\usage{
plotly_stackedtrace_kk_fbm_fbmw(
  drmin,
  drmax,
  betaF,
  betat,
  betaphi,
  kappa,
  nc,
  dr0 = 1,
  nround = 2,
  xlim = NULL,
  ylim = NULL,
  show_legend = T,
  hoverinfo_classes = "skip"
)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaF}{Load sharing coefficient (numeric array)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{kappa}{Weibull survival function shape parameter (numeric scalar)}

\item{nc}{The number of discrete, equal-width root classes to use}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}

\item{nround}{number of decimals used in rounding numbers shown in plot
(default 2)}

\item{xlim}{optional max limit for x-axis}

\item{ylim}{optional max limit for y-axis}

\item{show_legend}{if TRUE, shows legend}

\item{hoverinfo_classes}{if TRUE, shows reinforcements for all root classes
in hoverinfo information. If `skip`, this information is not shown and
only information for the total reinforcement is displaced in the
hoverinfo}
}
\value{
Plotly object
}
\description{
Creates a plotly object showing how reinforcement is mobilised according
to the FBM model with a discrete number of equal-width root diameter
classes. Power-law relations for the root area ratio and tensile strength
are assumed. The results are normalised by the Wu/Waldron (WWM) solution
using the same classes.
}
\examples{
plotly_stackedtrace_kk_fbm_fbmw(1, 5, 1, -0.5, -0.2, 4, 5)
}
