% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_phir.R
\name{predict_phir}
\alias{predict_phir}
\title{Predict cumulative root area ratio traces based on fitted power-law values}
\usage{
predict_phir(
  drmin,
  drmax,
  phirt,
  betaphi,
  dr = NULL,
  n = 101,
  cumulative = TRUE
)
}
\arguments{
\item{drmin}{the minimum root diameter (numerical scalar)}

\item{drmax}{the maximum root diameter (numerical scalar)}

\item{phirt}{the total root area ratio}

\item{betaphi}{the power-law coefficient in the root diameter
versus root area ratio fit}

\item{dr}{if specified, the traces are calculated at these
diameters. If not specified (default), a range is taken between
`drmin` and `drmax` using `nc` points.}

\item{n}{number of points used to plot the trace between
`drmin` <= dr <= `drmax` (used if `nc` is not specified)}

\item{cumulative}{if TRUE, cumulative trace is returned.
if FALSE, the distribution of root area ratio over all
root diameters is returned (i.e. the derivative of the c
cumulative trace)}
}
\value{
dataframe with predictions. Contained fields for
  the root diameter (`dr`).
  the cumulative root area ratio (`phir_cum`), if `cumulative==T`.
  the distribution curve (`phir`), if `cumulative==F`.
}
\description{
This function generates a root-diameter versus cumulative root area ratio
trace, given the inputs from a (bounded) power law fit between the root
diameter and cumulative root area ratio. These can be obtained using the
function `fit_powerlaw_phir`.
}
\examples{
predict_phir(1, 5, 0.01, -0.5, cumulative = TRUE)
predict_phir(1, 5, 0.01, -0.5, cumulative = FALSE)
predict_phir(1, 5, 0.01, -0.5, dr = seq(0, 10, l=25), cumulative = TRUE)
}
